move fields to audit into their own apps
