+++ apps/experiments/model_audit_fields.py
@@ -1,5 +1,3 @@
-from django.contrib.auth.models import AbstractUser
-
 PROMPT_FIELDS = ["owner", "name", "description", "prompt", "input_formatter", "team"]
 
 EXPERIMENT_FIELDS = [
@@ -36,8 +34,6 @@ CONSENT_FORM_FIELDS = [
     "team",
 ]
 
-TEAM_FIELDS = ["name", "slug", "members"]
-MEMBERSHIP_FIELDS = ["team", "user", "role"]
 EXPERIMENT_CHANNEL_FIELDS = [
     "name",
     "experiment",
@@ -48,9 +44,3 @@ EXPERIMENT_CHANNEL_FIELDS = [
 ]
 
 NO_ACTIVITY_CONFIG_FIELDS = ["message_for_bot", "name", "max_pings", "ping_after", "team"]
-MESSAGING_PROVIDER_FIELDS = ["type", "name", "team"]
-VOICE_PROVIDER_FIELDS = ["type", "name", "team"]
-LLM_PROVIDER_FIELDS = ["team", "type", "name", "llm_models"]
-
-# The auditing library struggles with dates. Let's ignore them for now
-CUSTOM_USER_FIELDS = [f.attname for f in AbstractUser._meta.fields if f.attname not in ["last_login", "date_joined"]]
