+++ apps/service_providers/model_audit_fields.py
@@ -1,0 +1,3 @@
+MESSAGING_PROVIDER_FIELDS = ["type", "name", "team"]
+VOICE_PROVIDER_FIELDS = ["type", "name", "team"]
+LLM_PROVIDER_FIELDS = ["team", "type", "name", "llm_models"]
