+++ apps/teams/model_audit_fields.py
@@ -1,0 +1,2 @@
+TEAM_FIELDS = ["name", "slug", "members"]
+MEMBERSHIP_FIELDS = ["team", "user", "role"]
